+++ changelog.md
@@ -5,7 +5,7 @@
 -   add `ScheduleQuery` class that provides three additional capabilities compared to `delayFactory`:
     -   `query.nextValidDate(date: Date)` – to find next closest date after a specific date (rather than delay)
     -   `query.next()` – gets closest date within provided schedule
-    -   `query.isValid(date = new Date())` – checks if provided date is within the schedule
+    -   `query.isValid(date)` – checks if provided date is within the schedule
 
 ### 0.3.0 → 0.3.1 (unstable)